Correccion de coliciones

+++ SpaceInvader/src/model/Alien.java
@@ -2,7 +2,7 @@ package model;
 
 public class Alien extends Character {
 
-	public Alien(String path, int x, int y, int lenght, int width, int deltaX, int deltaY) {
+	public Alien(String path, int x, int y, int lenght, int width, double deltaX, double deltaY) {
 		super(path, x, y, lenght, width, deltaX, deltaY);
 	}
